Using resource queue to provision apatelets

+++ pkg/scenario/deserialize/yaml_test.go
@@ -51,6 +51,6 @@ tasks:
 
 	assert.Equal(t, "testnode", obtainedScenario.Nodes[0].NodeType)
 	assert.Equal(t, "2G", obtainedScenario.Nodes[0].RAM)
-	assert.Equal(t, int32(42), obtainedScenario.Nodes[0].CPU)
-	assert.Equal(t, int32(42), obtainedScenario.Nodes[0].MaxPods)
+	assert.Equal(t, int64(42), obtainedScenario.Nodes[0].CPU)
+	assert.Equal(t, int64(42), obtainedScenario.Nodes[0].MaxPods)
 }
